Delete users (#24)
* added .python-version to gitignore

* Updated delete_user in users_service to delete the entire partition
corresponding to a user email and also added a delete scan abstract
method to find all entries with the substr in the sk and delete all
those entries. A basic test of the scan_sk_for_substr worked on the test
db, so the rest should work assuming the delete_partition method works
and delete_item method works. Also made an honest effort on error
handling

* Extended the db methods to include a scan by sk for a substr. Extended
this method to delete results found in the scan. Also improved error
handling for delete_item and delete_partition

* made changes according to PR feedback. Mostly logging and error handling
related

* added postman collection to mock events to test delete users endpoint